Day 4, focus Grid on removals
We only care about the rolls we can remove, so the Grid is intialised
with a VecDeque of rolls that can be removed (part 1 solution), which
is updated as each roll is removed (part 2 solution).

There is no need to store a set or map of locations where a roll is
present, so no hashing cost needs to be paid.